==> 3. post_annotation_generate_clean.ipynb <==
import pandas as pd
import glob
from scipy.stats import pearsonr
from scipy import stats
import numpy as np

class HIT(object):
    def __init__(self, row, df_s):

        self.hit_id = row['HITId']
        self.worker_id = row['WorkerId']
        self.assignment_id = row['AssignmentId']
        self.working_time = round(row['WorkTimeInSeconds']/60,2)
        self.ans = row['Answer.Q1']
        self.status = row['AssignmentStatus']
        
        self.quality_score = 0
        self.annotation_res = {}
        for data in row['Answer.Q1'].split('|'):
            id,_,ans = data.split('__')
            if 'TRAP-SAME' in id:
                if int(ans) >= 90:
                    self.quality_score += 1
            elif 'TRAP-DIFFERENT' in id:
                if int(ans) <= 10:
                    self.quality_score += 1
            else:
                self.annotation_res[id] = int(ans)
        self.task_id = -1
        for idx, datum in df_s.iterrows():
            if datum['text_0'].replace('\'','') in self.ans:
                self.task_id = idx
                break

        # comment this line if you dont want to apply z-normalization
        self.normalize() 

    def normalize(self):
        keys = list(self.annotation_res.keys())
        scores = list(self.annotation_res.values())
        zscore = stats.zscore(scores)
        ret = {}
        for idx in range(len(keys)):
            ret[keys[idx]] = zscore[idx]
        self.annotation_res = ret
        
def calculate_correlation(data):
    v1=[]; v2=[]; v3=[]
    for key in data[0].annotation_res.keys():
        v1.append(data[0].annotation_res[key])
        v2.append(data[1].annotation_res[key])
        v3.append(data[2].annotation_res[key])
    v1 = np.array(v1)
    v2 = np.array(v2)
    v3 = np.array(v3)
    pearson1, _ = pearsonr(v1, (v2+v3)/2)
    pearson2, _ = pearsonr(v2, (v1+v3)/2)
    pearson3, _ = pearsonr(v3, (v1+v2)/2)
    return (pearson1+pearson2+pearson3)/3

LANG='ES'
path = f'mturk/annotation_result/{LANG}/clean/*' #all downloaded mturk batch files
df_s = pd.read_csv(f'mturk/images/{LANG}/turker.csv')

hits = []
for file in glob.glob(path):
    df = pd.read_csv(file)
    for idx, row in df.iterrows():
        hits.append(HIT(row, df_s))

cols = ['task_id', 'hit_id', 'worker_id', 'assignment_id', 'status', 'quality_score', 'working_time(min)', 'annotation_res']
res = pd.DataFrame(columns=cols)
for idx in range(len(hits)):
    res = res.append({
            'task_id': hits[idx].task_id,
            'hit_id': hits[idx].hit_id,
            'worker_id': hits[idx].worker_id,
            'assignment_id': hits[idx].assignment_id,
            'status': hits[idx].status,
            'working_time(min)': hits[idx].working_time,
            'quality_score': hits[idx].quality_score,
            'annotation_res': hits[idx].annotation_res
        }, ignore_index=True)

res =res[res['quality_score']>=7]

res = res.sort_values(by=['task_id'], ignore_index=True)
res.to_csv(f'mturk/annotation_result/{LANG}/analysis_clean.csv', columns=cols, index=False)

data = {}
for id, row in res.iterrows():
    idx = row['task_id']
    if data.get(idx,[]) == []:
        data[idx] = [row]
    else:
        data[idx].append(row)

corrs = {}
for key in data.keys():
    assert len(data[key])==3
    correlation = calculate_correlation(data[key])
    corrs[key] = correlation
print(corrs)
print(np.mean(list(corrs.values())))

res

## Merge Human Annotation

import pandas as pd
from ast import literal_eval
from scipy import stats
from scipy.stats import pearsonr
LANG='ES'

source = f'mturk/annotation_result/{LANG}/analysis_clean.csv'
df = pd.read_csv(source)

data = {}
for idx, row in df.iterrows():
    id = row['task_id']
    if data.get(id, None) == None:
        data[id] = [row['annotation_res']]
    else:
        data[id].append(row['annotation_res'])

merge = []
for key in data.keys():
    a = literal_eval(data[key][0])
    b = literal_eval(data[key][1])
    c = literal_eval(data[key][2])
    for id in a.keys():
        overall = (a[id]+b[id]+c[id])/3
        merge.append((id, overall))
print(len(merge))

focus = []
coverage = []
for ids, score in merge:
    data_name, data_id, a, b, model_name = ids.split('-')
    if a == 'SYSTEM' and b == 'REFERENCE': #coverage
        coverage.append((data_id,data_name,model_name,score))
    elif a == 'REFERENCE' and b == 'SYSTEM': #focus
        focus.append((data_id,data_name,model_name,score))
print('focus:', len(focus))
print('coverage:', len(coverage))

df_focus = pd.DataFrame(columns = ['id', 'model', 'score'])
df_coverage = pd.DataFrame(columns = ['id', 'model', 'score'])

for data_id,_,model_name,score in focus:
    df_focus = df_focus.append(
        {'id': data_id, 'model': model_name, 'score': score},
        ignore_index=True
    )
for data_id,_,model_name,score in coverage:
    df_coverage = df_coverage.append(
        {'id': data_id, 'model': model_name, 'score': score},
        ignore_index=True
    )

df_focus.to_csv(f'mturk/annotation_result/{LANG}/human_focus_final.csv', index=False)
df_coverage.to_csv(f'mturk/annotation_result/{LANG}/human_coverage_final.csv', index=False)
