==> 5. transform_RU_and_ZH.ipynb <==
## Russian

import json, pickle, argparse
import glob, os


def write(fname, string):
    file = open(fname,'w')
    file.write(string.lower())
    file.close()

def get_dic(str1, str2, str3):
    dic = {}
    id=0
    for word in (' '.join(str1)).split(' ') + (' '.join(str2)).split(' ') + (' '.join(str3)).split(' '):
        word = word.strip()
        if dic.get(word,-1)==-1:
            dic[word] = id
            id+=1
    return dic

def convert(dic, list_str):
    new_list = []
    for string in list_str:
        new_str = []
        for word in string.split(' '):
            new_str.append(f'w{dic[word.strip()]}')
        new_list.append(' '.join(new_str))
    return new_list

def fix(sents):
    return [s.replace(' [UNK] ', '') for s in sents]

source = 'resulting_data/RU'
target = 'resulting_data/RU_latin'
os.makedirs(target+'/gold', exist_ok=True)
os.makedirs(target+'/pred_BERT', exist_ok=True)
os.makedirs(target+'/pred_PG', exist_ok=True)

for file in glob.glob(source+'/gold/*'):
    fname = file.split('/')[-1]
    gold = open(file).readlines()
    sys_BERT = fix(open(source+'/pred_BERT/'+fname).readlines())
    sys_PG = fix(open(source+'/pred_PG/'+fname).readlines())
    
    dic = get_dic(gold, sys_BERT, sys_PG)
    
    gold = convert(dic, gold)
    sys_BERT = convert(dic, sys_BERT)
    sys_PG = convert(dic, sys_PG)
    
    write(target+'/gold/'+fname, '\n'.join(gold))
    write(target+'/pred_BERT/'+fname, '\n'.join(sys_BERT))
    write(target+'/pred_PG/'+fname, '\n'.join(sys_PG))

## Chinese

def write(fname, string):
    file = open(fname,'w')
    file.write(string.lower())
    file.close()

def get_dic(str1, str2, str3):
    dic = {}
    id=0
    for word in (' '.join(str1)).split(' ') + (' '.join(str2)).split(' ') + (' '.join(str3)).split(' '):
        word = word.strip()
        if dic.get(word,-1)==-1:
            dic[word] = id
            id+=1
    return dic

def convert(dic, list_str):
    new_list = []
    for string in list_str:
        new_str = []
        for word in string.split(' '):
            new_str.append(f'w{dic[word.strip()]}')
        new_list.append(' '.join(new_str))
    return new_list

def fix(sents):
    temp = ''.join([s.replace('[UNK]', '') for s in sents])
    return temp.replace(' ','')

source = 'resulting_data/ZH'
target = 'resulting_data/ZH_latin'
os.makedirs(target+'/gold', exist_ok=True)
os.makedirs(target+'/pred_BERT', exist_ok=True)
os.makedirs(target+'/pred_PG', exist_ok=True)

for file in glob.glob(source+'/gold/*'):
    fname = file.split('/')[-1]
    gold = ''.join(open(file).readlines()).replace('\n','')
    sys_BERT = fix(open(source+'/pred_BERT/'+fname).readlines())
    sys_PG = fix(open(source+'/pred_PG/'+fname).readlines())
    
    dic = get_dic(gold, sys_BERT, sys_PG)
    
    gold = convert(dic, gold)
    sys_BERT = convert(dic, sys_BERT)
    sys_PG = convert(dic, sys_PG)
    
    write(target+'/gold/'+fname, ' '.join(gold))
    write(target+'/pred_BERT/'+fname, ' '.join(sys_BERT))
    write(target+'/pred_PG/'+fname, ' '.join(sys_PG))
