==> 6. correlation_of_traditional_metrics_combined.ipynb <==
import pandas as pd
import numpy as np
from scipy.stats import pearsonr 

LANG = 'FR'
MAIN = f'resulting_data/{LANG}_latin/'
MAIN = f'resulting_data/{LANG}/'

path_trad_BERT = MAIN+'BERT_traditional.csv'
path_trad_PG = MAIN+'PG_traditional.csv'

# BERT dan PG
path_human_focus = f'mturk/annotation_result/{LANG}/human_focus_final.csv'
path_human_coverage = f'mturk/annotation_result/{LANG}/human_coverage_final.csv'

def eval (h, c):
    for col in c.columns[1:]:
        score = pearsonr(h, c[col])[0]
        print(col+','+str(score))
    print()

trad_PG = pd.read_csv(path_trad_PG)
trad_BERT = pd.read_csv(path_trad_BERT)

human_foc = pd.read_csv(path_human_focus)
human_cov = pd.read_csv(path_human_coverage)

#focus combined
filtered_human = []
filtered_trad = pd.DataFrame(columns=trad_PG.columns)
for idx, row in human_foc.iterrows():
    filtered_human.append(row['score'])
    if row['model'] == 'BERT':
        filtered_trad = filtered_trad.append(trad_BERT[trad_BERT['id']==row['id']])
    elif row['model'] == 'PG':
        filtered_trad = filtered_trad.append(trad_PG[trad_PG['id']==row['id']])
print('Focus', MAIN, len(filtered_human))
eval(filtered_human, filtered_trad)

#coveraged combined
filtered_human = []
filtered_trad = pd.DataFrame(columns=trad_PG.columns)
for idx, row in human_cov.iterrows():
    filtered_human.append(row['score'])
    if row['model'] == 'BERT':
        filtered_trad = filtered_trad.append(trad_BERT[trad_BERT['id']==row['id']])
    elif row['model'] == 'PG':
        filtered_trad = filtered_trad.append(trad_PG[trad_PG['id']==row['id']])
print('Coverage', MAIN, len(filtered_human))
eval(filtered_human, filtered_trad)
