==> 7. correlation_of_bert_scores_combined.ipynb <==
%matplotlib inline
import pandas as pd
from scipy.stats import pearsonr, spearmanr
import matplotlib.pyplot as plt
plt.style.use('seaborn-whitegrid')
import numpy as np

lang2mono = {'EN': 'bert-base-uncased',
             'ID': 'indobert-base-uncased',
             'FR': 'camembert-base',
             'TR': 'bert-base-turkish-uncased',
             'ZH': 'bert-base-chinese',
             'RU': 'rubert-base-cased',
             'DE': 'bert-base-german-dbmdz-uncased',
             'ES': 'bert-base-spanish-wwm-uncased'}

LANGS = ['EN', 'ID', 'FR', 'TR', 'ZH', 'RU', 'DE', 'ES']

model2layer = {}
model2layer ['focus']= {
    'bert-base-multilingual-cased': 12,
    'bert-base-multilingual-uncased': 12,
    'xlm-roberta-base': 4,
    'xlm-roberta-large': 10,
    lang2mono['EN']: 1,
    lang2mono['ID']: 2,
    lang2mono['FR']: 10,
    lang2mono['TR']: 12,
    lang2mono['ZH']: 8,
    lang2mono['RU']: 4,
    lang2mono['DE']: 12,
    lang2mono['ES']: 4
}
model2layer ['coverage']= {
    'bert-base-multilingual-cased': 5,
    'bert-base-multilingual-uncased': 6,
    'xlm-roberta-base': 4,
    'xlm-roberta-large': 9,
    lang2mono['EN']: 2,
    lang2mono['ID']: 2,
    lang2mono['FR']: 9,
    lang2mono['TR']: 4,
    lang2mono['ZH']: 9,
    lang2mono['RU']: 12,
    lang2mono['DE']: 12,
    lang2mono['ES']: 4
}

def read_array(arr):
    arr = arr.replace('[','').replace(']','')
    num = arr.split(', ')
    num = [float(a) for a in num]
    return num

def read_array2(arr):
    arr = arr.replace('[','').replace(']','')
    num = arr.split(', ')
    num = [a.replace('\'','') for a in num]
    return num

def align(score, doc_id, human_score, model):
    human = []
    machine = []
    for idx, doc in enumerate(doc_id):
        human.append(human_score[(human_score['model'] == model) & (human_score['id']==int(doc))]['score'].values[0])
        machine.append(score[idx])
    return human, machine


def read_human_annotation(lang, types='focus'):
    if types=='focus':
        path_human = f'mturk/annotation_result/{lang}/human_focus_final.csv'
    else:
        assert types=='coverage'
        path_human = f'mturk/annotation_result/{lang}/human_coverage_final.csv'
    human = pd.read_csv(path_human)
    return human

def read(lang, human_score, prec_or_rec, pretrained):
    path_BERT = f'bert_score/{lang}--BERT--{pretrained}.csv'
    path_PG = f'bert_score/{lang}--PG--{pretrained}.csv'
    
    if prec_or_rec == 'precision':
        layer = model2layer['focus'][pretrained]-1
    elif prec_or_rec == 'recall':
        layer = model2layer['coverage'][pretrained]-1
    
    humans = []; machines = []

    row_BERT = pd.read_csv(path_BERT).iloc[layer]
    score = read_array(row_BERT[prec_or_rec])
    doc_id = read_array2(row_BERT['fnames'])
    human, machine = align(score, doc_id, human_score, 'BERT')
    humans += human
    machines += machine
    
    row_PG = pd.read_csv(path_PG).iloc[layer]
    score = read_array(row_PG[prec_or_rec])
    doc_id = read_array2(row_PG['fnames'])
    human, machine = align(score, doc_id, human_score, 'PG')
    humans += human
    machines += machine    
    
    #print(humans)
    #print(machines)
    #return spearmanr(humans, machines)[0], len(humans)
    return pearsonr(humans, machines)[0], len(humans)


LANGS = ['EN', 'ID', 'FR', 'TR', 'ZH', 'RU', 'DE', 'ES']

# Focus
print('Focus\n======')
cors = {}
for lang in LANGS:
    human_score = read_human_annotation(lang, 'focus')    
    for pretrained in [ lang2mono[lang], 'bert-base-multilingual-cased', 'bert-base-multilingual-uncased', \
                      'xlm-roberta-base', 'xlm-roberta-large']:     
        cor, num = read(lang, human_score, 'precision', pretrained)
        print(lang, num, pretrained, cor)
    print()

# Coverage
print('Coverage\n========')
pretrained = 'bert-base-multilingual-cased'
cors = {}
for lang in LANGS:
    human_score = read_human_annotation(lang, 'coverage')
    for pretrained in [ lang2mono[lang], 'bert-base-multilingual-cased', 'bert-base-multilingual-uncased', \
                      'xlm-roberta-base', 'xlm-roberta-large']:  
        cor, num = read(lang, human_score, 'recall', pretrained)
        print(lang, num, pretrained, cor)
    print()
